consolidation of similar files
