--- frontend/src/lib/components/AdminNavigation.svelte
@@ -1,9 +1,0 @@
-<nav class="list-nav p-4">
-	<ul>
-		<li><a href="/admin/create">Create User / Admin</a></li>
-		<li><a href="/admin/delete">Delete User</a></li>
-		<li><a href="/admin/update">Update User</a></li>
-		<li><a href="/admin/find">Find User</a></li>
-		<li><a href="/admin/skelly">Skelly</a></li>
-	</ul>
-</nav>
