--- frontend/src/lib/components/ClientNavigation.svelte
@@ -1,8 +1,0 @@
-<nav class="list-nav p-4">
-	<ul>
-		<li><a href="/client/upload-csv-test">Upload CSV Test</a></li>
-		<li><a href="/client/accounts">Accounts</a></li>
-		<li><a href="/client/budget/mine">Budget</a></li>
-		<li><a href="/client/skelly">Skelly</a></li>
-	</ul>
-</nav>
